requirements.txt
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Django==2.2.16
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.2
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyMySQL==1.0.2
python-decouple==3.1
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9